pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "reddit-survey-scheduler"
version = "0.1.0"
description = "Human-approved scheduler for posting survey recruitment links to an allowlist of Reddit communities"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "praw>=7.7",
]

[project.scripts]
reddit-scheduler = "reddit_scheduler.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
